+++ ChangeLog.rst
@@ -1,3 +1,15 @@
+Release 1.1.0 - 2018/08/27
+--------------------------
+
+* Changes to Supported Systems
+    * Ubuntu 18.04 LTS is now officially supported (#1230, #1231, #1236)
+
+    * Improvements
+        * Use double instead of float (#1232)
+            
+            * Bug fixes
+                * Fix the all-in-packaging tool to work on RHEL (#1235)
+
 Release 1.0.9 - 2018/04/23
 --------------------------